refactor: substituir double por BigDecimal para cálculos financeiros

+++ src/main/java/br/com/itau/geradornotafiscal/model/Item.java
@@ -6,6 +6,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @NoArgsConstructor
 @AllArgsConstructor
 @Setter
@@ -18,13 +20,10 @@ public class Item {
 	    private String descricao;
 
 	    @JsonProperty("valor_unitario")
-	    private double valorUnitario;
+	    private BigDecimal valorUnitario;
 
 	    @JsonProperty("quantidade")
 	    private int quantidade;
-
-
-
 
 }
